+++ src/mb/Mb_v06.hpp
@@ -174,6 +174,9 @@ struct ModelItem : BrowserListItem {
 		if (!moduleWidget) return;
 		APP->scene->rack->addModuleAtMouse(moduleWidget);
 
+		// Load template preset
+		moduleWidget->loadTemplate();
+
 		// Push ModuleAdd history action
 		history::ModuleAdd* h = new history::ModuleAdd;
 		h->name = "create module";